Renamed `vecmath` module to `math`
See https://github.com/PistonDevelopers/graphics/issues/884
