+++ src/ellipse.rs
@@ -7,7 +7,7 @@ use internal;
 use triangulation;
 use Graphics;
 use Color;
-use vecmath::Matrix2d;
+use math::Matrix2d;
 use DrawState;
 
 /// Ellipse border
